add interaction when installing

+++ install.sh
@@ -1,10 +1,21 @@
 #!/bin/bash
 
 cp ./bash_aliases ~/.bash_aliases
-echo "install ~/.bash_aliases"
+echo "1. install ~/.bash_aliases"
 
 cp ./bash_aliases.android ~/.bash_aliases.android
-echo "install ~/.bash_aliases.android"
+echo "2. install ~/.bash_aliases.android"
 
-cp ./bash_aliases.local ~/.bash_aliases.local
-echo "install ~/.bash_aliases.local"
+if [[ -f ~/.bash_aliases.local ]]; then
+    read -p "~/.bash_aliases.local exists! Do you want to overwrite it? [Y|n] : " answer
+    if [[ ${answer} == 'Y' ]]; then
+        cp ./bash_aliases.local ~/.bash_aliases.local
+        echo "3. update ~/.bash_aliases.local"
+        echo "install OK!"
+    elif [[ ${answer} == 'n' ]]; then
+        echo "install OK!"
+    else
+        echo "unknown choice"
+        echo "install OK!"
+    fi
+fi
